test: split terminal runtime fakes out of the workbench harness
